requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
attrs==24.3.0
cffi==1.17.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
flatbuffers==25.1.21
fonttools==4.55.5
fsspec==2024.12.0
hydra-core==1.3.2
iopath==0.1.10
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mediapipe==0.10.20
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
portalocker==3.1.1
protobuf==4.25.5
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
-e git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4#egg=SAM_2
sam2==1.1.0
scipy==1.15.1
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sounddevice==0.5.1
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
